fix struct processing in old dynamic components
Because of an oversight, some old dynamic components can currently miss
types or handle them multiple times, due to the way the removal of the
registry type signal was accommodated. Until they are migrated to the
updated design, this can be fixed by re-introducing the deprecated type
signal to the registry.