overhaul channel and connection error handling
Exception handling in channels and connections is crucial to ensure
proper cleanup and prevent undefined behavior when connections are
closed expectedly or unexpectedly. In general, exceptions can occur
under the following circumstances:

1. When transmitting instances
2. During asynchronous receive handling
3. When asynchronously transmitting instances to groups
4. In error handlers themselves

Additionally, these situations are not mutually exclusive. For example,
a host transceiver can receive an instance from a connection, which
then might break during group distribution to itself or other guest
connections.

Currently there is no comprehensive approach to error handling in
connections and channels, which leads, among others, to the following
issues:

* Undefined behavior when an exception occurs during transmission of
instances in the receive handler of a connection, which currently does
not interrupt the control flow and thus does not prevent repeated
access on already broken connections
* Undefined behavior when connections were pushed to the event queue
for removal but have incoming instances before them
* Undefined behavior when an exception occurs during the transmission
of a connection's sign-off message in its destructor
* Undefined behavior when a channel is destroyed while there is still
a corresponding async receive in the event queue
* Unexpected application termination when an exception occurs in an
error or transition handler callback

This commit introduces an overhauled approach to error handling in
channels and connections, which consists of the following:

1. All exceptions that occur during async receive handling are
propagated to the channel which then handles them via its error handler
2. If an exception occurs during a publish (synchronously or
asynchronously), it is passed to the error handler of the connection
3. Error and transition handlers are implemented as de facto noexcept,
even if this can currently not be enforced, due to restrictions of
std::function
4. Connections are closed when handling an error and are removed
immediately in a transceiver's transition handler
5. Async receive implementations in channels can now use weak_from_this
to check whether the channel has already been destroyed when the
callback is invoked