fix merge operation for dynamic struct types
The merge operation on two dynamic struct types currently fails, when
at least one of the included properties is not a substruct. This is
because the implementation uses the assignment operator to set the
value of the left-hand side property. While this works on static
properties, for proxy properties an assignment leads to the assignment
of the proxy and not the value itself.

This is a design flaw that should be solved, because it is unintuitive
and can lead to silent errors, as seen here.