make double conversion of duration explicit
A duration can currently be implicitly constructed and converted from
and to a double. This can be error-prone, because it may not be obvious
to every user that the duration is being represented by a floating
point second value including a fraction. Adding to this confusion is
the fact that a duration is currently sometimes specified by an integer
literal instead of an explicit double value.

Instead, the user should be encouraged to use chrono literals instead,
which makes creation of duration values much more expressive and at the
same time less error-prone as well.