adjust duration conversions for consistency
A duration can currently be converted to seconds, milliseconds and
microseconds. While the latter two return a converted integer value,
the first one actually refers to the count of the duration and returns
a double. To reduce the potential for confusion and the need for
additional conversions when a rounded second value is required, these
functions should provide a consistent behavior for acquiring rounded
integer values.