overhaul string conversion of chrono components
String conversion is currently implemented by using POSIX-exclusive
functions such as strptime, localtime_r, as well as the TZ environment
variable mechanism for distinguishing between local and UTC time.

Not only is the implementation not portable to non-POSIX platforms, it
also does not produce a standardized output. Timestamps are projected
onto their epoch durations, which in turn get converted to a ISO 8106
like format. However, timestamps can currently only be converted to
strings in their local timezone, with a fixed comma to separate the
part of the fractional seconds. Conversely, timestamps can currently
not be constructed from valid UTC timestamps.

Additionally, durations can only be converted to and from timestamps or
raw count values, while steady time points can only be converted like
system time points, even though they should rather be represented as
durations as well..

This commit overhauls the string conversion by utilizing the date
library by Howard Hinnant, which was, with small changes, adopted as a
specification for the upcoming ISO C++20 standard. It provides portable
support for dates and timezones in a way that is compatible to
std::chrono.

Using this library, string conversions are changed as follows:

- Duration objects are converted to and from ISO 8106 compatible
duration representations
- System time points are per default converted to and from ISO 8106
date times, including arbitrary time zones
- Steady time points are per default converted to and from ISO 8106
compatible durations
- All time points can be converted to and from other formats supported
by the date library

Note that the library is currently only used internally and therefore
does not have to be exposed as an additional dependency to the user.