make uuid's pure values
The uniqueness of unique id's also implies an immutability after they
have been created. If a uuid can have a default zero value and can
change after creation, a user always has to worry about the possibility
of the value not being initialized.

Instead, it is semantically more clear, if a uuid is always valid after
creation and cannot change unless overwritten by another instance. This
also encourages users to use a std::optional should an uninitialized
variable be required.