replace standalone ASIO with Boost version v1.69
The motivation to use the standalone version of ASIO instead of the
version from the Boost library was to resolve version dependency issues
between Boost and CMake (see original commit
017eddfb2f1a9e43b41e234d3c37d852ccc2a4f3).

However, since version v1.70, Boost has proper config support for CMake
projects and recent versions can be used easily without having to
update the version of CMake. This means that the standalone version of
ASIO can be exchanged with the Boost version again.

Since the ASIO version from Boost >= v1.70 has a few breaking changes,
this commit only upgrades to Boost v1.69, which contains the same ASIO
version as v1.12.2; the one reference in the ASIO standalone submodule.