upgrade Boost to version 1.72.0
Note that using the websocketpp library with a Boost version >= 1.70.0
requires using a revision from the development branch, since the master
branch has not been updated for roughly two years. This should be
considered only a preliminary solution until the WebSocket components
have been migrated to using the Beast from the Boost library.