replace Websocketpp with Boost Beast in core library
The WebSocket components in the DOTS core library currently depend on
the Websocketpp library, which is imported as a submodule. This library
has not been properly maintained for about two years and the official
release is already incompatible with the currently required version of
Boost ASIO, which is why a revision of the development branch has to be
used.

This commit circumvents these issues by using the Beast library to
implement the WebSocket components. Not only is Beast a more modern
library that has a better integration with Boost ASIO, but because it
is a Boost library as well, it is also always guaranteed to be
compatible with the ASIO version used by the project.