remove harmful conversion op from any struct type
The AnyStruct currently has user-defined conversion operator templates
that allow conversions to static struct types. They are marked as
explicit with the intention to prevent unintentional implicit
conversions.

However, explicit conversions can actually also be used by the compiler
in the context of direct-initialization. This is an issue, because
modern emplace functions, such as emplace_back on std::vector, perform
direct-initialization without it being obvious to the user. This makes
it currently possible to add incompatible AnyStruct instances to a
vector of a particular static struct type, leading to undefined
behavior.

Instead, a user should be forced to always explicitly specify when a
AnyStruct should be converted to a particular static struct type.