restrict channels to be created as shared ptr only
The channel base class derives from std::enable_shared_from_this to
support lifetime checks in callback functions.

However, this requirement is not conveyed to the user and since there
are currently no safety mechanisms to ensure that channels are actually
created within shared pointers, they can easily be created on the stack
(e.g. as a member variable). This leads to certain asynchronous channel
callbacks not working or failing with unintuitive exceptions.

Instead, it should be enforced on a language level that channels are
only created within memory managed by shared pointers. To do this, this
commit introduces a shared_ptr_only helper class, that requires a key
object for its creation, which can only be accessed by a corresponding
factory function.