deprecate namespace and destination group
The namespace and destination group properties of the transport header
are currently only necessary to retain backwards compatibility with the
legacy implementation. Instead of having to handle them in various
locations, they can also just be handled in the TCP channel, which is
the only interface between the legacy and the current implementation.

Namespaces and groups can be re-implemented when the protocol is
overhauled.

In addition, the destination client id can also deprecated, since it is
used neither by the legacy nor by the overhauled implementation.