remove proxy prefix from property iterator
Property iterators can be used to iterate over arbitrary properties and
are just using proxy properties as a means of providing access to the
individual property elements.

Since they are therefore not actually iterating over proxy properties
themselves, it is a little bit of a misnomer to include that in the
names of the corresponding components.