replace isFromMyself flag with header property
The dedicated isFromMyself flag was introduced in the past, so that the
header of a message in some situations did not need to be mutable,
which allowed to avoid unnecessary copies. However, since the
introduction of transmissions, the header can and is being modified
when a transmission is received.

By using the corresponding header property again, the flag parameter
can be removed for simplicity.