move issuing of subscriptions to transceiver
For historical reasons, subscriptions are currently always associated
with a dispatcher instance and the descriptor of a particular DOTS
struct type. This is because the dispatcher used to be the single class
issuing subscriptions. Since there are now other types of
subscriptions, such as the ones for type descriptors themselves, which
cannot be managed by the dispatcher, the responsibility of issuing
subscriptions should rather lie in the transceiver than in the
dispatcher.

This commit decouples subscriptions from the dispatcher and type
descriptors entirely and replaces the fixed mechanism with a
user-defined unsubscribe handler, which is called when the subscription
goes out of scope. This makes it possible to use subscriptions in a
flexible way, without being bound to a specific variant of
subscription. Additionally, subscriptions will now exclusively be
issued by the transceiver.