optimize binary size of static descriptor accessor
The accessor for static descriptors is currently more complicated than
necessary and leads to a significant bloat in binary size. Since every
static type requires its own accessors, the bloat also scales with the
number of static types.

This commit simplifies the static accessors by avoiding a local static
variable and an additional template instantiation when the descriptor
is added to the static descriptor map.