make static descriptor operations optional
Since static types have to define a descriptor, all corresponding type
operations have to always be included in the binary, regardless of
whether they are actually being used or not. Even though explicit
operations are usually faster, they lead to quite a lot of bloat in the
binary, especially for struct types, where several non-trivial
operations are added for every additional type. Because certain areas
of application are more sensitive to binary size, using the static
operations within the descriptor definitions should be optional.

This commit introduces a mechanism that allows static types to specify
whether static descriptor operations should be used or not.

Note that the mechanism does not avoid the static operations completely
if they are disabled, but rather relays them to the base class, which
either leads to them being small or completely optimized away by the
compiler.

Finally, this change has no affect on static type operations (e.g. of
struct) when the type is used directly and is only relevant in untyped
contexts.