disable static struct descriptor ops as a default
Since the typeless struct operations are usually fast enough and most
applications are not performance critical, it makes more sense to
disable static struct descriptor operations as a default.

Additionally, this commit introduces a macro mechanism, which allows to
not only enable static struct descriptor operations globally, but also
on a per-type basis.