optimize descriptor map emplace for binary size
The function for emplacing a descriptor to the static descriptor map is
currently defined in the header and is usually inlined in the
corresponding descriptor initialization calls. This leads to quite a
lot of bloat in the binary, because initialization has to be done for
every static type.

Since emplacing a descriptor is usually only done once during
initialization, the benefits of inlining are minor and the binary size
can be reduced by moving the definition to a dedicated translation
unit.