fix short-circuiting issue in SFINAE expressions
For SFINAE to work, the substitution failure must be dependent on the
template parameters in one form or another. Particularly, when using a
regular logical conjunction in the condition of an std::enable_if, all
arguments must be dependent on the template parameters. Otherwise, the
compiler may use short-circuiting to conclude a substitution failure
that does not depend on the template parameters.

To fix this, the logical operations can be replaced by the C++17 ones
from the STL, which were introduced for this very purpose.