make code generator usage platform independent
Usage of the DOTS code generator is currently platform dependent in
multiple ways:

 - The generation command is invoked via a shell call

 - Using the code generator requires setting multiple values to the
 PYTHONPATH environment variable

 - It is assumed that Python 3 is installed and is associated with
 Python script files (':' vs ';' path separators)

Additionally, the current command handling is hard to read and there is
no error checking when conditions are not met or the generator call
fails.

This commit overhauls the usage of the code generator by using CMake
mechanisms to find the required Python version and starting Python with
the generator directly, without the need for a shell call.

Note that this requires a corresponding change in the DOTS code
generator.