resolve overload resolution ambiguity
If a member function from a base class is hidden by a using
declaration, it is also brought into the scope of the sub class. This
can introduce an ambiguity if the declaring class also has a function
that cannot be distinguished from the one in the base class (e.g.
because it only differs in return types).

Additionally, hiding a base class function in this scenario is
unnecessary anyway, because the overload resolution always prefers the
member from the immediate class to the one from the base.