avoid virtual calls in dynamic properties
To determine whether a static descriptor can be used, properties
currently check if the value type is either typeless or a dynamic type.

However, all dynamic types now also have a corresponding static type
that is managing the abstraction to the underlying dynamic type. A
dynamic struct is dynamic, but is itself a static type that can be used
via the static type system.

Since the corresponding (static) descriptors also only relay the
operations to those types, they should be used directly rather than
going through a virtual function call.