introduce dynamic enum type
Dynamic enum types are currently represented as int32. This leads to
ambiguity in certain scenarios and is also inconsistent with other
dynamic types (e.g. dynamic structs).

By introducing a dedicated type, dynamic enums can be represented
directly, which also allows for a dedicated descriptor that can be
explicitly marked as dynamic.