remove composite type from static descriptor tests
The unit tests for the static descriptor currently include a composite
type in addition to the fundamental types. Because this type is
classified as a struct type, it gets added to the static type system as
such, even though it is missing a proper struct descriptor. Depending
on the compiler, the type can be present as a static type in the
registry and interfere with other tests that expect a struct type to
have a struct descriptor.

Since proper unit tests for structs exist in other test suites anyway,
the composite parts in the tests can just be removed entirely, with
only a small loss in test coverage.