fix non-standard descriptor init in static structs
Static structs currently use a static data member to hold a descriptor
reference for quick access. The reference is initialized via a static
call to the descriptor type.

However, since the descriptor can only be declared after the struct
type, it is not yet available when the static struct is declared. This
can lead to a build error on some compilers.

To fix this, the initialization can referred to a static member
function, which in turn may then access the descriptor, as its
definition is not evaluated during declaration.