fix static initialization of dynamic types
Because non-dynamic type descriptors must be accessible directly, every
descriptor has a static function that provides access to the global
descriptor instance. This is also true for descriptors of dynamic types
(e.g. dynamic structs), which are as well represented by an explicit
non-dynamic type. However, because their descriptor is not
default-constructible, there can be no global instance and the
corresponding accessor throws an exception instead.

Additionally, to ensure that every non-dynamic type is added to the
descriptor map, even when the descriptor is not accessed explicitly,
every non-dynamic descriptor type has a static helper reference that is
initialized with the descriptor instance. This means every dynamic type
descriptor should currently throw an exception during static
initialization.

The reason this was not observable as an issue, is that some compilers
(e.g. GCC) detect that the static data member is not accessible and
remove the initialization entirely for both non-dynamic and dynamic
types. This can be tested by making it accessible, in which case
applications that use dynamic types crash with an exception on start.

Since there does not seem to be a mechanism that reliably achieves the
desired goal cross-platform, it should be removed entirely at the cost
of a minor performance impact in scenarios that are not that common
anyway.