make property offset calculation portable
The current offset calculation is based on the assumption that
properties are placed directly behind the property area of the static
struct base class. This means that struct types are tightly packed and
properties might be placed in the alignment padding of the static
struct base class if they are of appropriate size. This assumption
holds true for current versions of the GCC and Clang.

However, other compilers, such as MSVC, use a different packing
approach, like aligning every declared sub-struct according to the
pointer size as a default. Not only that, but MSVC actually allows the
configuration of struct member alignment (which also affects
sub-structs) to be specified via a compiler option (see [1]).

To make the offset calculation portable, this commit introduces a
dedicated property offset for static properties, that determines the
actual initial offset at compile time.

References

 - [1]
 https://docs.microsoft.com/en-us/cpp/build/reference/zp-struct-member-alignment?view=msvc-160