add support for Windows and Visual C++ / MSVC
This commit adds support for the Visual C++ (or MSVC) compiler from
Microsoft and for targeting Windows platforms. Right now, the support
will only be maintained for Visual C++ 2019 targeting Windows 10. Older
MSVC and Windows versions might work but are not officially supported.

Note that the DOTS timepoint types have different base classes compared
to the non-MSVC version. This is because the MSVC STL disallows custom
duration types as template arguments for timepoints, due to a different
interpretation of the ISO C++ standard specification.

Additionally, this commit also updates cbor-cpp for MSVC support and
googletest to solve a compiler error with MSVC in one of the unit
tests.