fix static library bundling
To ease packaging, the DOTS library is bundled with all its
dependencies into a library bundle, which is done by a dedicated bundle
target in CMake. This mechanism is currently broken in several ways.

When using GCC or Clang, the bundle target overwrites the non-bundled
library, which might happen while it is already being linked to other
targets in the build tree, leading to linker errors in some situations.

With MSVC, the library is not overwritten at all and because the
bundled dependencies are removed from the library target, other targets
from the build tree cannot properly link the library.

Instead of using a dedicated bundling target, this commit performs the
bundling as a post build step for the actual library target. Because it
is part of the target, the bundling is guaranteed to be finished when
other targets are starting to being linked. Additionally, the
dependencies are guarded by a build interface generator expression,
which causes them to not be added as dependencies to the target in the
install interface, which avoids the dependencies having to be removed.