fix shared library install on DLL platforms
While on non-DLL platforms CMake treats shared libraries as LIBRARY
targets, on DLL platforms (i.e. Windows-based systems including Cygwin)
the DLL part of a shared library is treated as a RUNTIME target (see
[1]).

This is because a on DLL platforms a shared library is often installed
next to executables.

References:

 - [1] https://cmake.org/cmake/help/v3.12/command/install.html