fix disconnects on redundant container removes
The container currently throws an exception when an instance is to be
removed that is not part of the container. Since this is not
safe-guarded during dispatch, the exception is passed to the caller. If
the caller is a transceiver, the connection will be closed due to
supposed misconduct.

However, redundant removes are normal in DOTS, since they can be
transmitted by several guests while not having been processed by the
host.

Instead of checking for the existence of the instance during dispatch,
the behavior should rather be implemented as part of the container.
This allows efficient instance removal without redundant lookups.