use STL durations in time points for all platforms
Since the MSVC STL disallows the usage of custom types as duration
argument for STL time points, the DOTS time points currently use STL
durations when the library is built with MSVC. This is not ideal,
because it causes the types to be different between compilers, which
might lead to users unintentionally writing non-portable code.

Considering that the only disadvantage of the approach used for MSVC is
to sometimes having to specify the DOTS duration type explicitly, it
would be simpler to always use it as the single general approach.