support property invalidation on container insert
Up to this point, properties once valid could not be invalidated by a
container update. However, since property invalidation is desirable in
many scenarios, it should be properly supported in DOTS.

With this commit, properties can be invalidated by including invalid
properties within the attributes property set of the DOTS header.