fix ASIO epoll disable in install interface
Usage of epoll by ASIO is currently disabled via the
BOOST_ASIO_DISABLE_EPOLL preprocessor definition. This definition is
exported publicly, so that users of the DOTS library also have epoll
disabled.

However, the export is currently guarded by an BUILD_INTERFACE
generator expression and is therefore not set for applications that are
linking the library from outside of the build tree.