avoid special meaning of full property sets
In the context of publishing instances, a property set with all its
values set is currently interpreted as the intention of wanting to
publish only the valid properties of the instance. This is potentially
confusing, especially since now also invalid properties can be
published. Also, in the case of a type with the maximal number of
properties, it would actually prevent the user entirely from publishing
an instance 'as is' including invalid properties.

Instead of giving certain property set values special meanings, the
parameters can just be made optional, which is a familiar mechanism for
many users to expect a default value. If a property set is provided, it
is always intersected with the property set of the type, which allows
the usage of the PropertySet::All constant to publish an instance 'as
is'.