remove nonsensical and superfluous leftover code
Note that according to the corresponding commit (see [1]), the
destruction of the sender property was added to reduce load, because
the guest can deduce the id itself if it is not valid, but was
unintentionally broken in a subsequent commit (see [2]).

However, this is actually an incompatible change for guests using the
legacy library and should therefore currently not be included anyway.

References:

 - [1] commit 0f628da187219447783f397b08bbbbd225afe55b: assume peer id on invalid sender id to reduce load

 - [2] commit af9e58446d79e69e2ccd0b56f6e47bf57f2227f1: include DOTS header in transmissions