use portable direct offsets for static properties
The variety of possible approaches and the fact that struct packing is
entirely implementation defined, makes it laborious and error-prone to
maintain the current approach for offset calculation; especially with
the desire to support multiple compilers and platforms.

A better solution would be to just determine the offsets directly. Even
though this is always possible, there currently does not seem to be a
way to accomplish this at compile time in a portable way.

However, both GCC and MSVC allow the direct calculation of offsets at
compile time via pointer arithmetic. Even though this approach is also
not portable in regards to compilers (e.g. Clang), it always works for
all platforms supported by GCC and MSVC.

Additionally, the portable runtime approach can be included as a
fallback, which can always be used at the cost of a little efficiency.