fix confusing type registry iterators
The type registry currently provides iterator functions, which are
relayed to the underlying descriptor map. However, the registry only
stores dynamic types in that descriptor map, as static (i.e.
non-dynamic) types are accessed via the static descriptor map. The
descriptors yielded by the iteration over the registry are therefore
not the same as the ones being used for lookup.

To fix this issue, this commit replaces the iterator functions with a
dedicated for-each member, that can be used to access the combined
dynamic and non-dynamic set of descriptors.