merge channel service into transceivers
The channel service was originally introduced to fit into the
corresponding concept of ASIO. ASIO services usually are supposed to
hold/manage the resources that are shared across an execution context
(see [1]). However, this is not the case for DOTS channels, which are
neither shared nor managed or owned by the channel service and are
rather part of a transceiver and its underlying components (e.g.
connections).

Instead of having to use the channel service as a cumbersome factory,
it should be possible to create channels and listeners directly where
they are needed (this currently mainly means transceivers). This also
reduces the chances of unintentionally mixing channels and listeners
from different execution contexts to be mixed up in the transceivers,
which is not supported and leads to undefined behavior.

References:

 - [1] https://www.boost.org/doc/libs/1_71_0/doc/html/boost_asio/reference/execution_context.html