allow static user types to be disabled in registry
In some scenarios, particularly for pure dynamic guests, it is
necessary to prevent static user types from being available in the
registry. This is especially important when static and dynamic
transceivers are created within the same application.

Note that this commit does not yet consider enum types in the
definition of user types, because it is currently not possible to
distinguish between system and user definitions of enums.