replace submodule with standalone code generator
Having the code generator as a Git submodule is a historical remnant,
because it used to not be available in any other way. But even when the
generator itself is part of the library, the user still has to make
sure that the system provides all the necessary dependencies.

However, since it is now available as a Python package, it is easier
for the library to just depend on the generator itself, which can
easily be installed along with its dependencies (e.g. via PIP).

Note that the code generator is currently still missing some versioning
information, which is necessary to ensure that the installed generator
is compatible with the library.