fix access violations in ASIO timers
Even though an instance of the Timer class cancels the underlying ASIO
timer on destruction, the timer callback can still be invoked, if it
has already been queued for execution (see [1]). If the user callback
captured references to any objects that were assumed to have the same
lifetime as the Timer instance, it might access resources that will
already have been destroyed.

This commit fixes the issue by introducing and additional lifetime
check for the timer callback.

References:

 - [1] https://www.boost.org/doc/libs/1_71_0/doc/html/boost_asio/reference/basic_waitable_timer/cancel/overload1.html