add string conversion for struct types
Because more elaborate use cases can always be covered by custom
converters or serializers, the default string conversion should
primarily be suited for more general scenarios where the user "just
needs to print a struct instance" (e.g. for log or debug output). The
conversions should therefore be easy to read in one line and formatted
in a way users are familiar with. Ideally, it should be brief and
similar to the initialization of struct instances.

A good fit for above criteria is the style of designated initializers.
Since designated initialization will replace the current property
initialization for DOTS struct types once C++20 is more widely
supported, it is better to use it already, instead of having to change
it later, once designated initialization becomes available. Especially
since property initializers are considerably more verbose.