add experimental designated struct init support
As a potential replacement for the current property initializers, this
commit adds experimental support for designated initialization of DOTS
struct types.

To make this possible, DOTS struct types have to satisfy the conditions
of aggregate types (see [1]).

The support needs to explicitly be enabled via a preprocessor
definition and requires a compatible C++20 capable compiler to work
(e.g. GCC >= 8, Clang >= 10 or MSVC >= 19.21, see [2]).

Note that one additional downside of the support is that property
construction has to be made publicly accessible, which increases the
risk of unintentionally constructing temporary out-of-struct
properties.

References:

 - [1] https://en.cppreference.com/w/cpp/language/aggregate_initialization

 - [2] https://en.cppreference.com/w/cpp/compiler_support