use specialization for typeless vector descriptor
By having the descriptor for typeless vectors as a dedicated class, it
is more cumbersome to use in generic contexts, because it has to be
handled as a special case compared to typed vector descriptors.

Instead, the typeless version should be descriptor specialization as it
is the case for all other types.

Note that the simplified name can still be kept as an alias to ease
usage in non-generic use cases.