allow UUIDs to be default constructible
Default construction of unique IDs has deliberately been disallowed up
to this point, because they can easily lead to situations where UUID
objects are created and forgotten to be initialized with actual unique
values.

However, in generic contexts it is beneficial to have all types be
default constructible. One example for this is deserialization, which
often requires a corresponding storage (i.e. buffer) to serialize into.