make property descriptor and iterators header-only
Profiling has shown that making the implementations header-only leads
to considerable improvements in performance with both GCC and MSVC.
Since most of the corresponding functions are trivial accessors or
comparatively small, they are usually good candidates for inlining.

Additionally, because the components are in many cases used in
conjunction with one another (e.g. during property iteration), there
seems to often be a high potential for additional optimizations.