introduce type visitor
A common operation on DOTS types is to traverse instances of structs,
properties and vectors (e.g. during various kinds of serialization).
Although this can partially be done by property iteration, there is
currently no dedicated component that supports recursive traversal
through arbitrary type structures without requiring a considerable
amount of redundant code.

The type visitor introduced with this commit is a base class for
implementing a variety of different components that are based on type
traversal.

It comes in two flavors:

For advanced usages, a derived visitor can be implemented as a derived
CRTP class, which retains all type information as long as possible and
supports completely statically resolved traversal if the type is known
at compile time. This is aimed at use cases where performance is
critical and a lot of flexibility and control over the traversal is
required.

For simpler use cases, there also is a typeless version of the visitor
base class, which can be implemented by using polymorphism.