introduce overhauled CBOR serializer
On the basis of the recently introduced serializer base, the CBOR
serialization can be overhauled by using the type visiting system.

To take full advantage of the static type unrolling, the new serializer
does not use an external library and instead implements the CBOR
encoding and decoding directly.

Benchmarking has shown, that depending on the compiler and the kind of
struct instances, the overhauled serializer is on average between 30%
and 60% faster than the current one. When the instance type is known at
compiler time, the speedup can increase up to well over 100%, reaching
400% in extreme cases.

Note that the overhauled serializer does not yet replace the legacy one
and currently has to be used explicitly.