introduce experimental CBOR serializer
In the current CBOR encoding, properties of struct instances are
serialized to arbitrarily ordered elements of a CBOR map.

Even though this allows for some flexibility, it also has the downside
of making deserialization more complex, as each property has to be
looked up in the corresponding struct descriptor.

This commit introduces an experimental CBOR serializer with an
alternative encoding. Instead of using CBOR maps, struct instances are
serialized as CBOR arrays with the valid property set as their first
element, followed by the property values in a well-defined order. By
avoiding the additional lookup, deserialization is simpler and more
efficient.

Depending on the input and compiler, the experimental serializer
provides an average 20% to 50% speedup in deserialization performance,
reaching 75% in some typed cases. Because tags are now serialized into
the initial valid property set value, serialization size goes down for
large instances, while being slightly higher for very small instances.

Note that the well-defined property order is currently the order of
definition from the corresponding model. Alternatively, they could also
be sorted by tag, which would allow short-circuiting on sparse
instances.