introduce base class for string serializers
Since the main purpose of string serialization is to be human-readable,
there is no "right" way of serializing struct instances, as different
preferences and areas of application demand dedicated styles and
formats. This can easily lead to multiple redundant implementations of
string serialization, as it is the case now.

To provide a common basis for future string serializers, this commit
introduces a base class template for string serializers with the goal
of making it simple to define a string serialization style without
having to re-implement things that are common to all string
serializers.

Instead of using polymorphism, the user only has to provide a traits
class to define style and format. The base class then automatically
defines the corresponding serialization and deserialization functions
at compile time. The implementation provides a default traits class
that uses the same style as the string serialization defined in the
descriptor classes.

If needed, individual serialization functions can be statically
overridden by using the CRTP mechanism of the type visitor.

The base class also takes care of whitespace formatting, such as single
or multi line output, as well as indentation, which can be configured
via a configuration class at runtime.

Note that for deserialization to work, the user has to make sure that
the style is not ambiguous.